EditOrderModal: prepare data before editing AO
